Update status code with more comments.
status can be cared by the homescreen before get it from the app.
In that case, the app should know how to handle the result.

If the homescreen already cares the result, the app doesn't need to
care it again (noramlly).

So the apps should know the result is cared or not.

Change-Id: Ib35fa9b18fa134402731fdd53860417d913e2019