Use the vconf for trying to make a connection.
If the shortcut service process fails to make a connection to the master,
the service process has to try to connect to it again.

But I don't use the timer.
Instead of the timer, use the vconf.
So only if the state of master process is changed,
the service process will try to make a connection.
Only if the state of the master is changed to "started".

Change-Id: Ic3e5fc5a690267377c5315b10224a3bd595320a4